# src/apidocs.json
{
    "openapi" : "3.0.0",
    "info": {
        "title": "API Warehouse Ban",
        "description": "API ini mengelola operasi dan pembatasan sistem gudang, termasuk fungsi untuk melarang dan mengizinkan kembali item atau pengguna tertentu dari gudang.",
        "version": "0.1.9"
    },
    "servers" : [
        {
            "url": "http://localhost:2000", 
            "description": "localhost"
        }
    ], 
...
    "paths": {
...
        "/barang/insert": {
            "post": {
                "tags": [
                    "Barang"
                ],
                "summary": "Post Barang",
                "description": "Menambahkan Data Barang",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "nama_barang": {
                                        "type": "string"
                                    },
                                    "kategori": {
                                        "type": "string"
                                    },
                                    "harga": {
                                        "type": "integer"
                                    },
                                    "stok": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sukses"
                    },
                    "400": {
                        "description": "Bad Request"
                    },

                    "404": {
                        "description": "not found"
                    },
                    "409": {
                        "description": "coflig"
                    }

                }
            }
        },
        "/barang/update/{id_barang}": {
            "patch": {
                "tags": [
                    "Barang"
                ],
                "summary": "Update Barang",
                "description": "Memperbarui Barang Berdasarkan ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id_barang",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Masukkan Id Barang"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "nama_barang": {
                                        "type": "string"
                                    },
                                    "kategori": {
                                        "type": "string"
                                    },
                                    "harga": {
                                        "type": "integer"
                                    },
                                    "stok": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sukses"
                    },
                    "400": {
                        "description": "Bad Request"
                    },

                    "404": {
                        "description": "not found"
                    },
                    "409": {
                        "description": "conflict"
                    }
                }
            }
        },
        "/barang/delete/{id_barang}": {
            "delete": {
                "tags": [
                    "Barang"
                ],
                "summary": "Delete Barang",
                "description": "Menghapus Data Barang",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id_barang",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Masukkan Id Barang"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sukses"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    }
}